Dont build an intermediate chunk, better code factoring
